requirements.txt
Flask==3.1.2
flask-cors==6.0.1
Flask-SocketIO==5.5.1
eventlet==0.40.3
psycopg2-binary==2.9.11